+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Breaking changes:
+- `position` structure changed from { bottom, right } to string ['left', 'right']. 'right' is default;
+- `baseUrl` renamed to `wsUrl`
+
+### Added:
+- responsive mobile layouts
 
 ## [0.1.0] - 24.06.2021